Pass precise position timestamp within `PlayerEvent` (#460)
The following `PlayerEvent::PositionChanged/SeekDone` events
should contains the precise position timestamps to pass them
to back to client, otherwise there are no possibility to
identify timestamp changes between two `PositionChanged` events
on the client side.

gst (5.104499333) -> servo (5)
gst (5.153333333) -> servo (5)

+++ examples/muted_player.rs
@@ -107,11 +107,11 @@ fn run_example(servo_media: Arc<ServoMedia>) {
             },
             PlayerEvent::VideoFrameUpdated => eprint!("."),
             PlayerEvent::PositionChanged(p) => {
-                if p == 2 && !muted {
+                if p as u64 == 2 && !muted {
                     println!("\nPosition is at 2sec, muting, 1 second of silence incoming");
                     servo_media.mute(&context_id, true);
                     muted = true;
-                } else if p == 3 && muted {
+                } else if p as u64 == 3 && muted {
                     println!("\nPosition is at 3sec, unmuting");
                     servo_media.mute(&context_id, false);
                     muted = false;
